changes to header.jsx for persistant breed state
